refactor(governance): mv proposal to governance mod
