+++ opentron/src/manager/controllers.rs
@@ -1,7 +1,6 @@
-use ::keys::Address;
+//! Fork controller.
+
 use constants::block_version::{BlockVersion, ForkPolicy};
-use log::{debug, info};
-use proto2::state::{proposal::State as ProposalState, Proposal};
 use state::keys;
 
 use super::Manager;
@@ -60,97 +59,5 @@ impl ForkController<'_> {
                 Ok(num_passed >= minimum_upgraded)
             }
         }
-    }
-}
-
-/// Proposal controller to handle proposals during maintenance.
-pub struct ProposalController<'m> {
-    manager: &'m mut Manager,
-}
-
-impl ProposalController<'_> {
-    pub fn new<'a>(manager: &'a mut Manager) -> ProposalController<'a> {
-        ProposalController { manager }
-    }
-
-    pub fn process_proposals(&mut self) -> Result<(), String> {
-        let latest_proposal_id = self.manager.state_db.must_get(&keys::DynamicProperty::LatestProposalId);
-        if latest_proposal_id == 0 {
-            debug!("no proposal yet");
-            return Ok(());
-        }
-
-        // NOTE: proposals are handled in reverse order
-        for proposal_id in (1..=latest_proposal_id).rev() {
-            let proposal = self.manager.state_db.must_get(&keys::Proposal(proposal_id));
-
-            if proposal.is_processed() {
-                debug!("proposal #{} is processed", proposal_id);
-                // NOTE: proposal number less than or equal to this is already processed.
-                return Ok(());
-            }
-
-            if proposal.is_cancelled() {
-                debug!("proposal #{} is cancelled", proposal_id);
-                continue;
-            }
-
-            let current_maintenance_time = self
-                .manager
-                .state_db
-                .must_get(&keys::DynamicProperty::NextMaintenanceTime);
-            if proposal.expiration_time <= current_maintenance_time {
-                info!("proposal #{} expired, counting votes of active witnesses...", proposal_id);
-                self.process_proposal(proposal)?;
-                continue;
-            }
-
-            debug!("proposal #{} is active", proposal_id);
-        }
-        Ok(())
-    }
-
-    fn process_proposal(&mut self, mut proposal: Proposal) -> Result<(), String> {
-        let active_witnesses = self.manager.get_active_witnesses();
-        if active_witnesses.len() != constants::MAX_NUM_OF_ACTIVE_WITNESSES {
-            info!("current number of active witnesses: {}", active_witnesses.len());
-        }
-        let approval_count = proposal
-            .approver_addresses
-            .iter()
-            .filter(|addr| active_witnesses.contains(Address::from_bytes(addr)))
-            .count();
-
-        // 70% approvals
-        if approval_count >= active_witnesses.len() * constants::SOLID_THRESHOLD_PERCENT / 100 {
-            info!(
-                "proposal #{} passed, parameters: {:?}",
-                proposal.proposal_id, proposal.parameters
-            );
-            // set dynamic parameters
-            for (&param, &value) in proposal.parameters.iter() {
-                self.manager
-                    .state_db
-                    .put_key(keys::ChainParameter::from_i32(param as i32).unwrap(), value)
-                    .map_err(|_| "db insert error")?;
-            }
-            proposal.state = ProposalState::Approved as i32;
-            self.manager
-                .state_db
-                .put_key(keys::Proposal(proposal.proposal_id), proposal)
-                .map_err(|_| "db insert error")?;
-        } else {
-            // disapprove
-            info!(
-                "proposasl #{} did not reach enough approval, disapproved",
-                proposal.proposal_id
-            );
-            proposal.state = ProposalState::Disapproved as i32;
-            self.manager
-                .state_db
-                .put_key(keys::Proposal(proposal.proposal_id), proposal)
-                .map_err(|_| "db insert error")?;
-        }
-        Ok(())
     }
 }
