+++ opentron/src/manager/governance/mod.rs
@@ -1,4 +1,5 @@
 //! Chain governance related.
 
 pub mod maintenance;
+pub mod proposal;
 pub mod reward;
